+++ cli/src/commands/test.js
@@ -2,8 +2,8 @@ const path = require('path')
 const { reporter } = require('@dhis2/cli-helpers-engine')
 const { runCLI } = require('@jest/core')
 const fs = require('fs-extra')
+const { loadEnvFiles } = require('../lib/env')
 const exitOnCatch = require('../lib/exitOnCatch')
-const loadEnvFiles = require('../lib/loadEnvFiles')
 const makePaths = require('../lib/paths')
 
 const getAppJestConfig = ({ jestConfigPath, paths }) => {